feat: audit hardening, secure credentials, and money in integer cents (1.7.0)
Security:
- Hash staff PINs with a per-user salt (db v6); legacy hashes verify and
  transparently upgrade on next successful login
- Store admin credentials via flutter_secure_storage on Android/iOS behind a
  CredentialStore abstraction (SharedPreferences fallback on desktop); minSdk 23
- Fix login lockout hang: LockState lives in auth state, an active lockout
  returns immediately, and the PIN tripping the 5-failure threshold raises its
  error on the triggering attempt

Money:
- Store all money as integer cents (db v7): products.price_cents,
  orders.total_cents, order_items.unit_price_cents, data-preserving rebuild
- Move models, repositories, dashboard/report charts, waiter flow, refunds and
  receipt templates to int cents; formatPriceCents is the canonical formatter
- Restore FK enforcement in onOpen (cannot toggle PRAGMA foreign_keys inside a
  sqflite upgrade transaction)

Also:
- Reorganize admin settings and shared widgets folders; misc audit fixes
- Bump to 1.7.0+15; sync Windows Runner.rc fallback; CHANGELOG entry
- Add audit report (report.md)
- flutter analyze clean, 93 tests pass

+++ lib/features/admin/providers/dashboard_provider.dart
@@ -6,10 +6,10 @@ import 'package:brewline/features/admin/providers/dashboard_period.dart';
 /// The headline numbers on the dashboard with their delta vs an equal-length
 /// previous window. Recomputes whenever the period or the journal changes.
 class PeriodKpis {
-  final double revenue;
+  final int revenue;
   final int orderCount;
   final int itemCount;
-  final double avgOrderValue;
+  final int avgOrderValue;
 
   /// Fractional change vs. the previous window (see [deltaPercent]).
   final double revenueDelta;
@@ -54,11 +54,11 @@ final dashboardKpisProvider = FutureProvider<PeriodKpis>((ref) async {
     orderCount: current.orderCount,
     itemCount: current.itemCount,
     avgOrderValue: current.avgOrderValue,
-    revenueDelta: deltaPercent(current.revenue, past.revenue),
+    revenueDelta: deltaPercent(current.revenue.toDouble(), past.revenue.toDouble()),
     orderDelta: deltaPercent(
       current.orderCount.toDouble(),
       past.orderCount.toDouble(),
     ),
-    avgDelta: deltaPercent(current.avgOrderValue, past.avgOrderValue),
+    avgDelta: deltaPercent(current.avgOrderValue.toDouble(), past.avgOrderValue.toDouble()),
   );
 });
